Rename file names to be consistent
