+++ src/components/EResourceSections/index.js
@@ -4,8 +4,8 @@ export { default as AgreementsList } from './AgreementsList';
 export { default as MonographResourceInfo } from './MonographResourceInfo';
 export { default as PackageContents } from './PackageContents';
 export { default as PackageInfo } from './PackageInfo';
-export { default as PackageContentItemCoverage } from './PackageContentItemCoverage';
-export { default as PackageContentItemInfo } from './PackageContentItemInfo';
+export { default as PCICoverage } from './PCICoverage';
+export { default as PCIInfo } from './PCIInfo';
 export { default as ParentPackageDetails } from './ParentPackageDetails';
 export { default as PCIFormCoverage } from './PCIFormCoverage';
 export { default as PCIFormInfo } from './PCIFormInfo';
